Attempting to fix USB transmit stall
The addition of the temperature messages has triggered some problem in
the USB driver. When I open the port to read, it hangs up my USB bus --
mouse stops working...I think I've getting a stream of "enter" commands
in my terminal, and I have to unplug the everything and plug it back in to
get it back.