Low gain sample delay and sync output updates
Adds a separate parameter for the sample delay when using low gain.
This is because larger capacitances take longer to charge.

Places rising and falling edge of scope sync output into atomic
section so that it can show the integration window on the scope.

Now a value of -1 in the scan sync parameter causes the sync pulse to
be generated for active capacitance measurement.