Add support for feedback controlled splitting (#2)
Add support for feedback control of electrodes by creating 5 configurable capacitance measurement groups, which are read at each 500Hz AC drive cycle, and two separate active electrode groups which can be driven at independently controlled duty cycles.

## Messaging Changes 

This modifying USB messaging and is not backwards compatible. The BulkCapacitance message is updated to include a group ID, the SetElectrodesMsg is modified to include a group ID and a setting. The FeedbackCommanddMsg is added to control the feedback settings. The DutyCycleUpdated message reports changes in duty cycle made by the controller.

## Other changes

- A bug in StaticCircularBuffer was fixed so that it handles being full correctly.
- Pin modm dependency on older version to avoid new C++20 requirement for now
- Fix backwards LE polarity
- Fix incorrect APB1/APB2 clock frequencies