Overhaul project to support ATSAM and STM32
This is a major restructuring to separate out common code (lib) and
create targets for both the original STM32 (stm32) and the new SAMG55
microcontrollers (sam).

USB stack is switched to the tinyusb driver -- supported by modm
now -- as it works for both MCUs.

A bootloader (dfu-sam) is added, to replace the DFU functionality which
is provided by the bootloader built into the stm32.

Because modm had no support for the SAMG55, this update depends on
updates submitted to the modm repository, and so pulls in the latest
modm commits.

This also includes experimental support for "auto sample delay" mode, in
which the duration of current integration is dynamically adjusted based
on current.